Use ProductAssociationType instead of extension
