+++ src/Sylius/Bundle/AdminBundle/Resources/config/services/form.xml
@@ -56,6 +56,11 @@
             <tag name="form.type_extension" />
         </service>
 
+        <service id="sylius.form.extension.type.product_image" class="Sylius\Bundle\AdminBundle\Form\Extension\ProductImageTypeExtension">
+            <argument>%sylius.model.product_variant.class%</argument>
+            <tag name="form.type_extension" />
+        </service>
+
         <service id="sylius_admin.form.extension.type.channel" class="Sylius\Bundle\AdminBundle\Form\Extension\ChannelTypeExtension">
             <tag name="form.type_extension" extended-type="Sylius\Bundle\ChannelBundle\Form\Type\ChannelType" priority="50" />
         </service>
