+++ src/Sylius/Bundle/AdminBundle/Resources/config/services/product/form.xml
@@ -17,13 +17,14 @@
             <tag name="form.type_extension" />
         </service>
 
-        <service id="sylius_admin.form.extension.type.product_associations" class="Sylius\Bundle\AdminBundle\Form\Extension\ProductAssociationsTypeExtension">
-            <tag name="form.type_extension" />
+        <service id="sylius_admin.form.type.product_associations" class="Sylius\Bundle\AdminBundle\Form\Type\ProductAssociationsType">
+            <argument type="service" id="sylius_admin.form.type.data_transformer.products_to_product_associations" />
+            <tag name="form.type" />
         </service>
 
-        <service id="sylius.form.extension.type.product_image" class="Sylius\Bundle\AdminBundle\Form\Extension\ProductImageTypeExtension">
-            <argument>%sylius.model.product_variant.class%</argument>
-            <tag name="form.type_extension" />
+        <service id="sylius_admin.form.type.data_transformer.products_to_product_associations" class="Sylius\Bundle\AdminBundle\Form\DataTransformer\ProductsToProductAssociationsTransformer">
+            <argument type="service" id="sylius.factory.product_association" />
+            <argument type="service" id="sylius.repository.product_association_type" />
         </service>
     </services>
 </container>
